requirements.txt
# 일반적인 학습을 위해 필요할 것으로 예상되는 Python 패키지를 정리함.
# Deploy에서는 불필요한 패키지를 제거하고 별도의 requirements.txt를 생성할 것.
# PyTorch 등 별도 설치 또는 빌드된 라이브러리는 별도 기재되지 않음.
# 원활한 업데이트 및 호환성을 위해 학습 단계에서는 모든 라이브러리를 기재하지 않으며
# version을 매우 엄격하게 설정하지는 않으나 대체로 가장 최신 버전을 사용할 것이라고 가정함.
# 다만, 필요에 따라 학습 중에도 버전을 고정하는 것을 허용함.
# 특히 PyTorch Lightning과 Monai의 버전은 학습 중에 고정하는 것이 용이함.

GPUtil
SimpleITK
accelerate
albumentations
black
cryptography
einops
fairscale
flake8
h5py
hydra-core
hydra-optuna-sweeper
indexed-gzip  # Faster decompression in Nibabel.
ipywidgets
iterative-stratification
joblib
jupyterlab
matplotlib
mkl
wandb
monai==1.0.*
nibabel
numpy
omegaconf
opencv-python-headless
openpyxl  # Read `.xlsx` files in Pandas.
pandas
pydicom
pylint
pymongo==3.10.1
pytest
pytorch-lightning==1.7.*
openmim
rich
scikit-image
scikit-learn
scipy
seaborn
tensorboard
torchinfo
tqdm
setuptools==59.5.0
timm
